+++ R/grouped_lm.R
@@ -36,10 +36,10 @@
 #' @seealso grouped_slr
 #'
 #' @examples
-#' 
+#'
 #' # loading needed libraries
 #' library(ggplot2)
-#' 
+#'
 #' # getting tidy output of results
 #' grouped_lm(
 #'   data = mtcars,
@@ -47,7 +47,7 @@
 #'   formula = mpg ~ am * wt,
 #'   output = "tidy"
 #' )
-#' 
+#'
 #' # getting model summaries
 #' # diamonds dataset from ggplot2
 #' grouped_lm(