+++ R/grouped_lmer.R
@@ -40,11 +40,11 @@
 #' @importFrom broom.mixed augment
 #'
 #' @examples
-#' 
+#'
 #' # loading libraries containing data
 #' library(ggplot2)
 #' library(gapminder)
-#' 
+#'
 #' # getting tidy output of results
 #' # let's use only 50% data to speed it up
 #' groupedstats::grouped_lmer(
@@ -53,7 +53,7 @@
 #'   grouping.vars = year,
 #'   output = "tidy"
 #' )
-#' 
+#'
 #' # getting model summaries
 #' # let's use only 50% data to speed it up
 #' grouped_lmer(