+++ R/grouped_summary.R
@@ -35,7 +35,7 @@
 #' @importFrom stats qt
 #'
 #' @examples
-#' 
+#'
 #' # another possibility
 #' groupedstats::grouped_summary(
 #'   data = datasets::iris,
@@ -43,7 +43,7 @@
 #'   measures = Sepal.Length:Petal.Width,
 #'   measures.type = "numeric"
 #' )
-#' 
+#'
 #' # if you have just one variable per argument, you need not use `c()`
 #' groupedstats::grouped_summary(
 #'   data = datasets::ToothGrowth,