+++ R/reexports.R
@@ -1,0 +1,13 @@
+# ----------------------- broomExtra -------------------------------------
+
+#' @export
+#' @importFrom broomExtra grouped_tidy
+broomExtra::grouped_tidy
+
+#' @export
+#' @importFrom broomExtra grouped_glance
+broomExtra::grouped_glance
+
+#' @export
+#' @importFrom broomExtra grouped_augment
+broomExtra::grouped_augment