+++ NEWS.md
@@ -3,7 +3,7 @@
 MAJOR CHANGES
 
   - New function `grouped_tidy` to extract a tidy dataframe of results in a
-    grouped analysis for any function.
+    grouped analysis for any function that has a `data` argument.
 
 # groupedstats 0.0.5