+++ NEWS.md
@@ -1,3 +1,5 @@
+# groupedstats 0.0.6.9000
+
 # groupedstats 0.0.6
  
 NEW FUNCTIONS